# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/automation/workflows/{id}":  {
                                                         "get":  {
                                                                     "operationId":  "AutomationController_findOne",
                                                                     "parameters":  [
                                                                                        {
                                                                                            "name":  "id",
                                                                                            "required":  true,
                                                                                            "in":  "path",
                                                                                            "schema":  {
                                                                                                           "type":  "string"
                                                                                                       }
                                                                                        }
                                                                                    ],
                                                                     "responses":  {
                                                                                       "200":  {
                                                                                                   "description":  ""
                                                                                               }
                                                                                   },
                                                                     "tags":  [
                                                                                  "Automation"
                                                                              ]
                                                                 },
                                                         "patch":  {
                                                                       "operationId":  "AutomationController_update",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "id",
                                                                                              "required":  true,
                                                                                              "in":  "path",
                                                                                              "schema":  {
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "requestBody":  {
                                                                                           "required":  true,
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/UpdateWorkflowDto"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       }
                                                                                       },
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  ""
                                                                                                 }
                                                                                     },
                                                                       "tags":  [
                                                                                    "Automation"
                                                                                ]
                                                                   }
                                                     },
                  "/api/automation/workflows/{id}/toggle":  {
                                                                "patch":  {
                                                                              "operationId":  "AutomationController_toggleActive",
                                                                              "parameters":  [
                                                                                                 {
                                                                                                     "name":  "id",
                                                                                                     "required":  true,
                                                                                                     "in":  "path",
                                                                                                     "schema":  {
                                                                                                                    "type":  "string"
                                                                                                                }
                                                                                                 }
                                                                                             ],
                                                                              "responses":  {
                                                                                                "200":  {
                                                                                                            "description":  ""
                                                                                                        }
                                                                                            },
                                                                              "tags":  [
                                                                                           "Automation"
                                                                                       ]
                                                                          }
                                                            },
...
                  "/api/automation/workflows/{id}/runs":  {
                                                              "get":  {
                                                                          "operationId":  "AutomationController_getRuns",
                                                                          "parameters":  [
                                                                                             {
                                                                                                 "name":  "id",
                                                                                                 "required":  true,
                                                                                                 "in":  "path",
                                                                                                 "schema":  {
                                                                                                                "type":  "string"
                                                                                                            }
                                                                                             }
                                                                                         ],
                                                                          "responses":  {
                                                                                            "200":  {
                                                                                                        "description":  ""
                                                                                                    }
                                                                                        },
                                                                          "tags":  [
                                                                                       "Automation"
                                                                                   ]
                                                                      }
                                                          },
...
              },
    "info":  {
                 "title":  "Trifusion-Dynamics Auth Service",
                 "description":  "The API description for Auth Service",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "UpdateWorkflowDto":  {
                                                                 "type":  "object",
                                                                 "properties":  {

                                                                                }
                                                             },
...
                                   }
                   }
}